+++ davaiops/__init__.py
@@ -1,5 +1,5 @@
 
-__version__ = '1.0.4'
-__update_date__ = '2022-06-25_18:09:50'
+__version__ = '1.0.5'
+__update_date__ = '2022-06-25_18:25:30'
 from pathlib import Path
 ROOT_PATH = Path(__file__)