+++ davaiops/routes/appgen.py
@@ -1,4 +1,7 @@
-from flask import Flask
+from flask import (
+    Flask,
+    url_for
+)
 from pukr import (
     get_logger,
     InterceptHandler
@@ -13,8 +16,8 @@ from davaiops.flask_base import (
 from davaiops.routes.admin import admin
 from davaiops.routes.api import api
 from davaiops.routes.koned import koned
-from davaiops.routes.errors import errors
 from davaiops.routes.main import main
+from davaiops.routes.projects import projects
 from davaiops.routes.user import users
 
 
@@ -37,7 +40,7 @@ def create_app(*args, **kwargs) -> Flask:
     app.extensions.setdefault('loguru', logger)
 
     # Register routes
-    for rt in [admin, api, koned, main, users, errors]:
+    for rt in [admin, api, koned, main, projects, users]:
         app.register_blueprint(rt)
 
     return app